Refactor to support native and service worker downloads.
Also implements support for skipping existing files and cleans up various
debugging statements.

Uses feature detection to determine native file API availability.
