+++ packages/openneuro-app/src/scripts/datalad/download/download-uri.js
@@ -1,0 +1,8 @@
+import config from '../../../../config.js'
+
+export const downloadUri = (datasetId, snapshotTag) =>
+  // This can't be a GraphQL query since it is intercepted
+  // by the service worker
+  snapshotTag
+    ? `${config.crn.url}datasets/${datasetId}/snapshots/${snapshotTag}/download`
+    : `${config.crn.url}datasets/${datasetId}/download`
