+++ packages/openneuro-app/src/scripts/datalad/download/track-download.js
@@ -1,0 +1,15 @@
+import { event } from 'react-ga'
+import datalad from '../../utils/datalad'
+
+export const trackDownload = (datasetId, snapshotTag) => {
+  event({
+    category: 'Download',
+    action: 'Started web download',
+    label: snapshotTag ? `${datasetId}:${snapshotTag}` : datasetId,
+  })
+  datalad.trackAnalytics(datasetId, {
+    snapshot: true,
+    tag: snapshotTag,
+    type: 'downloads',
+  })
+}
