+++ packages/openneuro-app/src/scripts/sw.js
@@ -26,19 +26,29 @@ self.addEventListener('fetch', event => {
   // for non-GET requests.
   if (event.request.method === 'GET') {
     const url = new URL(event.request.url)
-    // Skip serving from cache for cross origin 'only-if-cached' requests
+    // Check for native filesystem support before registering the custom download handler
     if (
-      event.request.cache === 'only-if-cached' &&
-      event.request.mode !== 'same-origin'
-    )
-      return
-    // Respond from cache, then the network
-    event.respondWith(
-      caches.open(CACHE_NAME).then(cache => {
-        return cache.match(event.request).then(response => {
-          return response || fetch(event.request)
-        })
-      }),
-    )
+      !('chooseFileSystemEntries' in window) &&
+      url.pathname.startsWith('/crn') &&
+      url.pathname.endsWith('download')
+    ) {
+      // Catch any aggregate download requests
+      return event.respondWith(bundleResponse(url))
+    } else {
+      // Skip serving from cache for cross origin 'only-if-cached' requests
+      if (
+        event.request.cache === 'only-if-cached' &&
+        event.request.mode !== 'same-origin'
+      )
+        return
+      // Respond from cache, then the network
+      event.respondWith(
+        caches.open(CACHE_NAME).then(cache => {
+          return cache.match(event.request).then(response => {
+            return response || fetch(event.request)
+          })
+        }),
+      )
+    }
   }
 })
